[CDC-63] Refacor performance tests
