+++ test/e2e/cdcSogeiTranscodeToken.js
@@ -5,6 +5,8 @@ import { loginFullUrl } from '../common/api/bpdIoLogin.js'
 import { assert, statusOk } from '../common/assertions.js'
 import { isEnvValid, isTestEnabledOnEnv, UAT } from '../common/envs.js'
 import dotenv from 'k6/x/dotenv'
+import { randomFiscalCode } from '../common/utils.js'
+
 
 const REGISTERED_ENVS = [UAT]
 
@@ -18,7 +20,7 @@ if (isEnvValid(__ENV.TARGET_ENV)) {
 }
 
 export function setup() {
-    const authToken = loginFullUrl(`${baseUrl}/bpd/pagopa/api/v1/login`, myEnv.FISCAL_CODE_EXISTING)
+    const authToken = loginFullUrl(`${baseUrl}/bpd/pagopa/api/v1/login`, randomFiscalCode())
     return {
         headers: {
             Authorization: `Bearer ${authToken}`,
